site/pyproject.toml
[project]
name = "allainclair-website"
version = "0.0.1"
description = "Allainclair Website: allainclair.com"
authors = [
	{name = "Allainclair Flausino dos Santos", email = "allainclair@gmail.com"},
]

requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
	"python-fasthtml>=0.12.29",
	"pydantic-settings>=2.11",
	"timeago>=1.0.16",
	"markdown-it-py>=4.0.0",
	"mdit-py-plugins>=0.5.0",
	"aiofiles>=25.1.0",
]

[dependency-groups]

dev = [
	"ruff",
	"mypy",
]
debug = [
	"rich",
	"ipython",
]


[tool.ruff]
line-length = 80

lint.select = [
	"A",
	"ARG",
	"ANN",  # Take care if there is conflict with mypy.
	"ASYNC",
	"B",
	"BLE",
	"COM",
	"C4",
	"C90",
	# "DTZ",  # https://docs.astral.sh/ruff/rules/#flake8-datetimez-dtz
	"E",
	"EM",
	"EXE",
	"F",
	"FBT",
	"FLY",
	"FURB",
	"G",
	"I",
	"ICN",
	"INP",
	"INT",
	"ISC",
	"LOG",
	"N",
	"PERF",
	"PGH",
	"PIE",
	"PL",
	"PT",
	"PTH",
	"PYI",
	"Q",
	"RET",
	"RSE",
	"RUF",
	"S",
	"SIM",
	"SLOT",
	"SLF",
	"T10",
	"T20",
	"TCH",
	"TD",
	"TID",
	"TRY",
	"UP",
	"W",
	"YTT"
]
lint.ignore = [
	"ANN401",  # https://docs.astral.sh/ruff/rules/any-type/

	# The following rules may cause conflicts when used with the formatter: `COM812`.
	"COM812",  # https://docs.astral.sh/ruff/rules/missing-trailing-comma/

	# The following rules may cause conflicts when used with the formatter: `ISC001`
	"ISC001",  # https://docs.astral.sh/ruff/rules/single-line-implicit-string-concatenation/

	# PLR2004 Magic value used in comparison
	"PLR2004",  # https://docs.astral.sh/ruff/rules/magic-value-comparison/

	"S101",  # https://docs.astral.sh/ruff/rules/assert/"
	"SIM108",  # https://docs.astral.sh/ruff/rules/if-else-block-instead-of-if-exp/

	"TD003", # https://docs.astral.sh/ruff/rules/missing-todo-link/
	"TD002", # https://docs.astral.sh/ruff/rules/missing-todo-author/

	# The `format.indent-style="tab"` option is incompatible with `W191`, which lints against all uses of tabs
	"W191",  # https://docs.astral.sh/ruff/rules/tab-indentation/
]

lint.mccabe.max-complexity = 5

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
line-ending = "lf"  # Use `\n` line endings for all files

[tool.ruff.lint.isort]
case-sensitive = true
